Initial Sketch of eliminating twice declared units
The problem:
    What happens if you declare units in the initial style and use
different units later?

To deal with this, this is an attempt at a solution that allows you to
declare units in the initStyle but not in the animation, meaning you
always have to use the units given in the init style.

However, there seems to be an issue in my approach :/