Compensate for large dt
if dt is very large (starting at, maybe, 300ms)
then it's most likely because someone left the
browser mid animation and then returned.
The browser 'pauses' the animation until it's viewed again.
The longer the user is gone, the longer the pause.
This can cause very screwy results, as you might imagine.
To fix this, if there is a large dt, then
* start is reset
* elapsed is reset
* this frame is essentially skipped.